refactoring: extracted Download from Config
